Add code that works for all test cases but 17 & 18
This code works for all the test cases except two. It creates a valid
BST, but one that needs a few optimizations.

The biggest problem at this point is that when a node is removed it's
left branch is inserted into it's right branch and the right branch
takes the place of the removed node.

This is not ideal as I have come to find out. Future updates should
correct this problem.